requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
pillow~=10.3.0
imutils~=0.5.4
tkinter